+++ wal-line-vc.el
@@ -1,4 +1,4 @@
-;;; wal-line-vc.el --- Indicate version control info -*- lexical-binding: t; -*-
+;;; wal-line-vc.el --- Version control segment -*- lexical-binding: t; -*-
 
 ;; Author: Krister Schuchardt <krister.schuchardt@gmail.com>
 ;; Homepage: https://github.com/Walheimat/wal-line
@@ -18,34 +18,22 @@
 (declare-function wal-line--spacer "wal-line.el")
 
 (defvar wal-line-vc--scope-regexp "\\(feature\\|\\(\\w+\\)?fix\\|improvement\\)\\/")
-(defvar-local wal-line-vc--info nil)
+
 (defvar-local wal-line-vc--state nil)
+(defvar-local wal-line-vc--info nil)
+(defvar-local wal-line-vc--segment nil)
 
-;; Taken from `doom-modeline'.
-(defun wal-line-vc--update-info (&rest _)
-  "Update the version control info."
-  (setq-local wal-line-vc--info
-        (when (and vc-mode buffer-file-name)
-          (let* ((backend (vc-backend buffer-file-name))
-                 (state (vc-state (file-local-name buffer-file-name) backend))
-                 (str (if vc-display-status
-                          (substring vc-mode (+ (if (eq backend 'Hg) 2 3) 2))
-                        "")))
-            (setq-local wal-line-vc--state state)
-            (propertize (replace-regexp-in-string
-                         wal-line-vc--scope-regexp
-                         ""
-                         str)
-                        'mouse-face 'wal-line-highlight
-                        'face (wal-line-vc--face-for-state))))))
+;; State:
 
-(defun wal-line-vc--segment ()
-  "Show version control info."
-  (unless wal-line-vc--info
-    (wal-line-vc--update-info))
-  (if wal-line-vc--info
-      (concat (wal-line--spacer) wal-line-vc--info)
-    ""))
+(defun wal-line-vc--update-state ()
+  "Update the version control state."
+  (when-let ((state (wal-line-vc--get-state)))
+    (setq-local wal-line-vc--state state)))
+
+(defun wal-line-vc--get-state ()
+  "Get the version control state."
+  (let ((backend (vc-backend buffer-file-name)))
+    (vc-state (file-local-name buffer-file-name) backend)))
 
 (defun wal-line-vc--face-for-state ()
   "Get the correct face for the state."
@@ -58,20 +46,48 @@
            'wal-line-contrast)
           (t 'wal-line-neutral))))
 
+;; Info:
+
+(defun wal-line-vc--update-info ()
+  "Update version control info."
+  (when-let ((info (wal-line-vc--get-info)))
+    (setq-local wal-line-vc--info info)))
+
+(defun wal-line-vc--get-info ()
+  "Get version control info."
+  (when (and vc-mode buffer-file-name)
+    (let* ((backend (vc-backend buffer-file-name))
+           (status (if vc-display-status
+                       (substring vc-mode (+ (if (eq backend 'Hg) 2 3) 2))
+                     ""))
+           (str (replace-regexp-in-string wal-line-vc--scope-regexp "" status)))
+      (propertize str
+                  'mouse-face 'wal-line-highlight
+                  'face (wal-line-vc--face-for-state)))))
+
+;; Segment:
+
+(defun wal-line-vc--update (&rest _)
+  "Update and (re-)set the segment."
+  (wal-line-vc--update-state)
+  (wal-line-vc--update-info)
+
+  (setq-local wal-line-vc--segment (concat (wal-line--spacer) wal-line-vc--info)))
+
 (defvar wal-line--segments)
 (wal-line-add-segment vc)
 
 (defun wal-line-vc--setup ()
   "Set up version control segment."
-  (add-hook 'find-file-hook #'wal-line-vc--update-info)
-  (add-hook 'after-save-hook #'wal-line-vc--update-info)
-  (advice-add #'vc-refresh-state :after #'wal-line-vc--update-info))
+  (add-hook 'find-file-hook #'wal-line-vc--update)
+  (add-hook 'after-save-hook #'wal-line-vc--update)
+  (advice-add #'vc-refresh-state :after #'wal-line-vc--update))
 
 (defun wal-line-vc--teardown ()
   "Tear down version control segment."
-  (remove-hook 'find-file-hook #'wal-line-vc--update-info)
-  (remove-hook 'after-save-hook #'wal-line-vc--update-info)
-  (advice-remove #'vc-refresh-state #'wal-line-vc--update-info))
+  (remove-hook 'find-file-hook #'wal-line-vc--update)
+  (remove-hook 'after-save-hook #'wal-line-vc--update)
+  (advice-remove #'vc-refresh-state #'wal-line-vc--update))
 
 (add-hook 'wal-line-setup-hook #'wal-line-vc--setup)
 (add-hook 'wal-line-teardown-hook #'wal-line-vc--teardown)